feat: implemented mapping for RocketLaunced message, test green

+++ dto/rockerLaunched.go
@@ -1,7 +1,7 @@
 package DTO
 
 type RocketLaunched struct {
-	LaunchSpeed int64  `json:"launchSpeed"`
+	LaunchSpeed int    `json:"launchSpeed"`
 	Mission     string `json:"mission"`
 	Type        string `json:"type"`
 }
